requirements.txt
aiodns==3.0.0
aiofile==3.7.4
aiofiles==0.7.0
aiogram==3.0.0b1
aiohttp==3.8.1
aiolimiter==1.0.0
aiosignal==1.2.0
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.25.0
attrs==21.4.0
autoflake==1.4
backports.zoneinfo==0.2.1
beanie==1.9.2
black==22.3.0
Brotli==1.0.9
cachetools==5.2.0
caio==0.9.3
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cycler==0.11.0
DAWG-Python==0.7.2
defusedxml==0.7.1
Django==4.0.6
django-autoslug==1.9.8
django-countries==7.3.2
django-environ==0.8.1
django-filter==21.1
django-phonenumber-field==6.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==1.16.0
docopt==0.6.2
dp==1.0
emoji==1.7.0
environs==9.5.0
flake8==4.0.1
Flask==2.1.2
fonttools==4.29.1
frozenlist==1.3.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
greenlet==1.1.2
gspread==5.4.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.3.2
loguru==0.6.0
magic-filter==1.0.6
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.14.1
matplotlib==3.5.1
mccabe==0.6.1
motor==2.5.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.22.2
oauth2client==4.1.3
oauthlib==3.2.0
openapi-python-client==0.11.4
opencv-python==4.5.5.62
packaging==21.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.47
Pillow==9.0.1
platformdirs==2.5.2
protobuf==4.21.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyDrive==1.3.1
pyflakes==2.4.0
PyJWT==2.4.0
pymongo==3.12.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-i18n==0.3.9
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.6.1
serializers==0.2.4
setuptools-scm==6.4.2
shellingham==1.4.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
SQLAlchemy==1.4.35
sqlparse==0.2.4
stripe==3.5.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
typer==0.4.2
typing_extensions==4.0.1
ujson==5.1.0
uritemplate==4.1.1
urllib3==1.26.10
uvloop==0.16.0
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
